refactor(executors): 收敛根目录入口
