+++ src/cli/lib/runtime-defaults.ts
@@ -1,8 +1,8 @@
 import { accessSync, constants } from 'node:fs';
 import { delimiter, join } from 'node:path';
 import { getCodexModelSuggestion } from './codex-model-hint.js';
-import { DEFAULT_MODEL, JUDGE_MODEL } from '../../executors/defaults.js';
-import { executorFamily } from '../../executors/registry.js';
+import { DEFAULT_MODEL, JUDGE_MODEL } from '../../executors/core/defaults.js';
+import { executorFamily } from '../../executors/core/registry.js';
 
 export interface RuntimeResolutionOptions {
   env?: NodeJS.ProcessEnv;
