+++ src/dsh-plugin/host-executor.ts
@@ -1,8 +1,8 @@
 import { randomUUID } from 'node:crypto';
 import type { ExecResult, ExecutorFn, ExecutorInput } from '../types/index.js';
 import { buildDshHostResult, type DshHostRunResult } from './protocol.js';
-import { createDshHostRuntimeFingerprint } from '../executors/runtime-fingerprint.js';
-import { DEFAULT_TIMEOUT_MS } from '../executors/defaults.js';
+import { createDshHostRuntimeFingerprint } from '../executors/core/runtime-fingerprint.js';
+import { DEFAULT_TIMEOUT_MS } from '../executors/core/defaults.js';
 
 type UnknownRecord = Record<string, unknown>;
